feat: Create MyProfileTab by reusing new ProfileLayout and navigate to  ProfileEditPage when pressing edit button
