+++ lib/ui/pages/home/home_page.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:lang_mate/ui/pages/home/tabs/chat/chat_tab.dart';
+import 'package:lang_mate/ui/pages/home/tabs/profile/my_profile_tab.dart';
 import 'package:lang_mate/ui/pages/home/tabs/users/matched_users_tab.dart';
 import 'package:lang_mate/ui/pages/home/widgets/home_bottom_navigation_bar.dart';
-import 'package:lang_mate/ui/pages/profile/user_profile_page.dart';
-import 'package:lang_mate/ui/user_global_view_model.dart';
 
 import '../profile_edit/profile_edit_page.dart';
 import 'home_view_model.dart';
@@ -24,7 +23,7 @@ class HomePage extends StatelessWidget {
             children: [
               MatchedUsersTab(),
               ChatTab(),
-              ProfileEditPage(user: ref.read(userGlobalViewModelProvider)!),
+              MyProfileTab(),
             ],
           );
         },
