+++ lib/ui/pages/profile_edit/profile_edit_page.dart
@@ -218,8 +218,8 @@ class _ProfileEditPageState extends ConsumerState<ProfileEditPage> {
                           vertical: 12,
                         ),
                         decoration: BoxDecoration(
-                          color: Colors.white,
                           borderRadius: BorderRadius.circular(10),
+                          color: Colors.white,
                           boxShadow: [
                             BoxShadow(
                               color: Colors.black.withValues(alpha: 0.1),
